removing file extension restriction
resolves #13

In order to do so, we have to have the ability to ignore certain files that were being generated by the .snag.yml (like the binary generated from 'go build').
ExcludeDirectory was therefore renamed to 'ignore' and behave simlarly to a .gitignore. You can list files or directories and both will be ignored