Parsing string args correctly
There was a bug in the code that cause commands that had quotes in them to not work as expected.
Commands are no longer split just on a whitespace. If an opening quote is found the split function
finds its ending quote and returns the entire chunk as one argument.